backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
anthropic==0.40.0
openpyxl==3.1.2
python-dotenv==1.0.0
